Making common sum implementation.
