+++ tiledb/sm/query/readers/aggregators/sum_aggregator.cc
@@ -45,52 +45,10 @@ class SumAggregatorStatusException : public StatusException {
   }
 };
 
-/** Specialization of safe_sum for int64_t sums. */
-template <>
-void safe_sum<int64_t>(int64_t value, int64_t& sum) {
-  if (sum > 0 && value > 0 &&
-      (sum > (std::numeric_limits<int64_t>::max() - value))) {
-    sum = std::numeric_limits<int64_t>::max();
-    throw std::overflow_error("overflow on sum");
-  }
-
-  if (sum < 0 && value < 0 &&
-      (sum < (std::numeric_limits<int64_t>::min() - value))) {
-    sum = std::numeric_limits<int64_t>::min();
-    throw std::overflow_error("overflow on sum");
-  }
-
-  sum += value;
-}
-
-/** Specialization of safe_sum for uint64_t sums. */
-template <>
-void safe_sum<uint64_t>(uint64_t value, uint64_t& sum) {
-  if (sum > (std::numeric_limits<uint64_t>::max() - value)) {
-    sum = std::numeric_limits<uint64_t>::max();
-    throw std::overflow_error("overflow on sum");
-  }
-
-  sum += value;
-}
-
-/** Specialization of safe_sum for double sums. */
-template <>
-void safe_sum<double>(double value, double& sum) {
-  if ((sum < 0.0) == (value < 0.0) &&
-      (std::abs(sum) >
-       (std::numeric_limits<double>::max() - std::abs(value)))) {
-    sum = sum < 0.0 ? std::numeric_limits<double>::lowest() :
-                      std::numeric_limits<double>::max();
-    throw std::overflow_error("overflow on sum");
-  }
-
-  sum += value;
-}
-
 template <typename T>
 SumAggregator<T>::SumAggregator(const FieldInfo field_info)
     : field_info_(field_info)
+    , summator_(field_info)
     , sum_(0)
     , validity_value_(
           field_info_.is_nullable_ ? std::make_optional(0) : nullopt)
@@ -154,47 +112,31 @@ void SumAggregator<T>::validate_output_buffer(
 
 template <typename T>
 void SumAggregator<T>::aggregate_data(AggregateBuffer& input_data) {
-  tuple<typename sum_type_data<T>::sum_type, optional<uint8_t>> res{0, nullopt};
+  // Return if a previous aggregation has overflowed.
+  if (sum_overflowed_) {
+    return;
+  }
 
-  bool overflow = false;
   try {
+    tuple<typename sum_type_data<T>::sum_type, uint64_t, optional<uint8_t>> res{
+        0, 0, nullopt};
+
     if (input_data.is_count_bitmap()) {
-      res = sum<typename sum_type_data<T>::sum_type, uint64_t>(input_data);
+      res =
+          summator_.template sum<typename sum_type_data<T>::sum_type, uint64_t>(
+              input_data);
     } else {
-      res = sum<typename sum_type_data<T>::sum_type, uint8_t>(input_data);
-    }
-  } catch (std::overflow_error&) {
-    overflow = true;
-  }
-
-  {
-    // This might be called on multiple threads, the final result stored in sum_
-    // should be computed in a thread safe manner. The mutex also protects
-    // sum_overflowed_ which indicates when the sum has overflowed.
-    std::unique_lock lock(sum_mtx_);
-
-    // A previous operation already overflowed the sum, return.
-    if (sum_overflowed_) {
-      return;
+      res =
+          summator_.template sum<typename sum_type_data<T>::sum_type, uint8_t>(
+              input_data);
     }
 
-    // If we have an overflow, signal it, else it's business as usual.
-    if (overflow) {
-      sum_overflowed_ = true;
-      sum_ = std::get<0>(res);
-      return;
-    } else {
-      // This sum might overflow as well.
-      try {
-        safe_sum(std::get<0>(res), sum_);
-      } catch (std::overflow_error&) {
-        sum_overflowed_ = true;
-      }
+    safe_sum(std::get<0>(res), sum_);
+    if (field_info_.is_nullable_ && std::get<2>(res).value() == 1) {
+      validity_value_ = 1;
     }
-  }
-
-  if (field_info_.is_nullable_ && std::get<1>(res).value() == 1) {
-    validity_value_ = 1;
+  } catch (std::overflow_error& e) {
+    sum_overflowed_ = true;
   }
 }
 
@@ -203,86 +145,30 @@ void SumAggregator<T>::copy_to_user_buffer(
     std::string output_field_name,
     std::unordered_map<std::string, QueryBuffer>& buffers) {
   auto& result_buffer = buffers[output_field_name];
-  *static_cast<typename sum_type_data<T>::sum_type*>(result_buffer.buffer_) =
-      sum_;
+  auto s =
+      static_cast<typename sum_type_data<T>::sum_type*>(result_buffer.buffer_);
+  if (sum_overflowed_) {
+    *s = std::numeric_limits<typename sum_type_data<T>::sum_type>::max();
+  } else {
+    *s = sum_;
+  }
 
   if (result_buffer.buffer_size_) {
     *result_buffer.buffer_size_ = sizeof(typename sum_type_data<T>::sum_type);
   }
 
   if (field_info_.is_nullable_) {
-    *static_cast<uint8_t*>(result_buffer.validity_vector_.buffer()) =
-        validity_value_.value();
+    auto v = static_cast<uint8_t*>(result_buffer.validity_vector_.buffer());
+    if (sum_overflowed_) {
+      *v = 0;
+    } else {
+      *v = validity_value_.value();
+    }
 
     if (result_buffer.validity_vector_.buffer_size()) {
       *result_buffer.validity_vector_.buffer_size() = 1;
     }
   }
-}
-
-template <typename T>
-template <typename SUM_T, typename BITMAP_T>
-tuple<SUM_T, optional<uint8_t>> SumAggregator<T>::sum(
-    AggregateBuffer& input_data) {
-  SUM_T sum{0};
-  optional<uint8_t> validity{nullopt};
-  auto values = input_data.fixed_data_as<T>();
-
-  // Run different loops for bitmap versus no bitmap and nullable versus non
-  // nullable. The bitmap tells us which cells was already filtered out by
-  // ranges or query conditions.
-  if (input_data.has_bitmap()) {
-    auto bitmap_values = input_data.bitmap_data_as<BITMAP_T>();
-
-    if (field_info_.is_nullable_) {
-      validity = 0;
-      auto validity_values = input_data.validity_data();
-
-      // Process for nullable sums with bitmap.
-      for (uint64_t c = input_data.min_cell(); c < input_data.max_cell(); c++) {
-        if (validity_values[c] != 0 && bitmap_values[c] != 0) {
-          validity = 1;
-
-          auto value = static_cast<SUM_T>(values[c]);
-          for (BITMAP_T i = 0; i < bitmap_values[c]; i++) {
-            safe_sum(value, sum);
-          }
-        }
-      }
-    } else {
-      // Process for non nullable sums with bitmap.
-      for (uint64_t c = input_data.min_cell(); c < input_data.max_cell(); c++) {
-        auto value = static_cast<SUM_T>(values[c]);
-
-        for (BITMAP_T i = 0; i < bitmap_values[c]; i++) {
-          safe_sum(value, sum);
-        }
-      }
-    }
-  } else {
-    if (field_info_.is_nullable_) {
-      validity = 0;
-      auto validity_values = input_data.validity_data();
-
-      // Process for nullable sums with no bitmap.
-      for (uint64_t c = input_data.min_cell(); c < input_data.max_cell(); c++) {
-        if (validity_values[c] != 0) {
-          validity = 1;
-
-          auto value = static_cast<SUM_T>(values[c]);
-          safe_sum(value, sum);
-        }
-      }
-    } else {
-      // Process for non nullable sums with no bitmap.
-      for (uint64_t c = input_data.min_cell(); c < input_data.max_cell(); c++) {
-        auto value = static_cast<SUM_T>(values[c]);
-        safe_sum(value, sum);
-      }
-    }
-  }
-
-  return {sum, validity};
 }
 
 // Explicit template instantiations
